# src/openapi.yaml
openapi: 3.1.0
info:
  title: Ping Pong API
  version: 0.1.0
paths:
...
  /matches/{matchId}:
    get:
      summary: Get match details
      parameters:
        - $ref: "#/components/parameters/matchId"
      responses:
        "200":
          description: Match details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetails"
...
components:
  parameters:
    matchId:
      in: path
      name: matchId
      required: true
      description: Unique identifier of the match.
      schema:
        type: string
        example: "match1"
...
  schemas:
    Side:
      type: string
      description: Which side of the table is currently serving or hitting.
      enum: [ping, pong]
    Score:
      type: object
      required: [ping, pong]
      properties:
        ping:
          type: integer
        pong:
          type: integer
    LongestRally:
      type: object
      description: Longest (by hit count) rally in the match. Duration included, but not a qualifying factor.
      required: [hitCount, duration]
      properties:
        hitCount:
          type: integer
        duration:
          type: string
          format: duration
    RallyState:
      type: object
      description: Current state of the ongoing rally.
      required: [side, hitCount]
      properties:
        side:
          $ref: "#/components/schemas/Side"
        hitTimeoutTimestamp:
          type: [string, "null"]
          format: date-time
          description: Time at which point will be lost if hit is not returned. Null if rally has not started yet.
        serveTimestamp:
          type: [string, "null"]
          format: date-time
          description: Time at which first hit of the rally was made. Null if rally has not started yet.
        hitCount:
          type: integer
          description: Number of hits from both sides since the start of the rally.
    GameState:
      type: object
      description: High-level game state, excluding current rally details.
      required: [server, score]
      properties:
        server:
          $ref: "#/components/schemas/Side"
        score:
          $ref: "#/components/schemas/Score"
        longestRally:
          oneOf:
            - $ref: "#/components/schemas/LongestRally"
            - type: "null"
    MatchDetails:
      type: object
      description: Complete details about a match.
      required: [rallyState, gameState]
      properties:
        rallyState:
          $ref: "#/components/schemas/RallyState"
        gameState:
          $ref: "#/components/schemas/GameState"